Improving formatting of table output
This is easier to parse while doing manual testing for now but this
should ideally be replaced with an algorithm for pretty printing the
table in columns (element, description). There also seems to be an issue
with simply calling .text() on a nested element from 'select' in that it
mashes things together in a unintuitive way (line breaks, not stripping
some tags/special chars but stripping others). Need to work on that as
well.